+++ web/src/pages/AdminPage.tsx
@@ -13,10 +13,15 @@ import {
 } from "react-bootstrap";
 import { useUser } from "../hooks/useUser";
 import { useToast } from "../hooks/useToast";
-import { addUserToStreamAlerts } from "../utils/apiService";
+import { useQuery, useQueryClient } from "@tanstack/react-query";
+import {
+  getStreamAlertsChannels,
+  addChannelToStreamAlerts,
+  removeChannelFromStreamAlerts,
+} from "../utils/apiService";
+import { TwitchUserData } from "../types/users";
 
 interface AdminPageProps {}
-
 const AdminPage: React.FunctionComponent<AdminPageProps> = () => {
   useEffect(() => {
     document.title = "Admin Panel | Helpasaur King";
@@ -24,8 +29,14 @@ const AdminPage: React.FunctionComponent<AdminPageProps> = () => {
 
   const toast = useToast();
   const { data: user, isLoading: userLoading } = useUser();
+  const { data: streamAlertsChannels } = useQuery({
+    queryKey: ["streamAlertsChannels"],
+    queryFn: getStreamAlertsChannels,
+  });
+  const queryClient = useQueryClient();
 
   const [userToAdd, setUserToAdd] = React.useState("");
+  const [userAddInProgress, setUserAddInProgress] = React.useState(false);
   const handleUserToAddInputChange = (
     e: React.ChangeEvent<HTMLInputElement>
   ) => {
@@ -33,13 +44,14 @@ const AdminPage: React.FunctionComponent<AdminPageProps> = () => {
     setUserToAdd(value);
   };
   const handleAddUserToStreamAlerts = async () => {
+    setUserAddInProgress(true);
     try {
-      const { data: userAddResult } = await addUserToStreamAlerts(userToAdd);
+      const { data: userAddResult } = await addChannelToStreamAlerts(userToAdd);
       const userResult = userAddResult[0].value;
-      console.log(userResult);
       if (userResult.status === "success") {
         toast.success(`Added ${userToAdd} to stream alerts!`);
         setUserToAdd("");
+        queryClient.invalidateQueries({ queryKey: ["streamAlertsChannels"] });
       } else if (userResult.status === "error") {
         toast.error(
           `Failed to add ${userToAdd} to stream alerts: ${userResult.message}`
@@ -48,6 +60,36 @@ const AdminPage: React.FunctionComponent<AdminPageProps> = () => {
     } catch (err) {
       toast.error(`Failed to add ${userToAdd} to stream alerts: ${err}`);
     }
+    setUserAddInProgress(false);
+  };
+
+  const [userToRemove, setUserToRemove] = React.useState("");
+  const [userRemoveInProgress, setUserRemoveInProgress] = React.useState(false);
+  const handleUserToRemoveInputChange = (
+    e: React.ChangeEvent<HTMLSelectElement>
+  ) => {
+    const value = e.target.value;
+    setUserToRemove(value);
+  };
+  const handleRemoveUserFromStreamAlerts = async () => {
+    setUserRemoveInProgress(true);
+    try {
+      const userResult = await removeChannelFromStreamAlerts(userToRemove);
+      if (userResult.result === "success") {
+        toast.success(`Removed ${userToRemove} from stream alerts!`);
+        setUserToRemove("");
+        queryClient.invalidateQueries({ queryKey: ["streamAlertsChannels"] });
+      } else if (userResult.result === "noop") {
+        toast.info(userResult.message);
+      } else if (userResult.result === "error") {
+        toast.error(userResult.message);
+      }
+    } catch (err) {
+      toast.error(
+        `Failed to remove ${userToRemove} from stream alerts: ${err}`
+      );
+    }
+    setUserRemoveInProgress(false);
   };
 
   if (userLoading)
@@ -115,13 +157,61 @@ const AdminPage: React.FunctionComponent<AdminPageProps> = () => {
             </FloatingLabel>
           </Col>
           <Col>
-            <Button
-              variant="dark"
-              onClick={handleAddUserToStreamAlerts}
-              size="lg"
+            {userAddInProgress ? (
+              <Spinner animation="border" role="statues">
+                <span className="visually-hidden">Loading...</span>
+              </Spinner>
+            ) : (
+              <Button
+                variant="dark"
+                onClick={handleAddUserToStreamAlerts}
+                size="lg"
+                disabled={userAddInProgress}
+              >
+                <i className="fa-regular fa-square-plus px-1"></i> Add User
+              </Button>
+            )}
+          </Col>
+        </Row>
+      </Container>
+
+      <Container>
+        <h2>Remove User from Stream Alerts</h2>
+        <Row>
+          <Col>
+            <FloatingLabel
+              controlId="userToRemove"
+              label="Choose a user to remove"
             >
-              <i className="fa-regular fa-square-plus px-1"></i> Add User
-            </Button>
+              <Form.Select
+                aria-label="Choose a user to remove"
+                value={userToRemove}
+                onChange={handleUserToRemoveInputChange}
+              >
+                <option>-</option>
+                {streamAlertsChannels?.map((channel: TwitchUserData) => (
+                  <option key={channel.id} value={channel.id}>
+                    {channel.display_name}
+                  </option>
+                ))}
+              </Form.Select>
+            </FloatingLabel>
+          </Col>
+          <Col>
+            {userAddInProgress ? (
+              <Spinner animation="border" role="statues">
+                <span className="visually-hidden">Loading...</span>
+              </Spinner>
+            ) : (
+              <Button
+                variant="dark"
+                onClick={handleRemoveUserFromStreamAlerts}
+                size="lg"
+                disabled={userRemoveInProgress}
+              >
+                <i className="fa-regular fa-square-minus px-1"></i> Remove User
+              </Button>
+            )}
           </Col>
         </Row>
       </Container>